Conflicts:
	src/main/java/org/interledger/ilp/InterledgerError.java
	src/main/java/org/interledger/ledger/model/AccountInfo.java

+++ src/main/java/org/interledger/codecs/oer/ilqp/QuoteByDestinationAmountResponseOerCodec.java
@@ -11,14 +11,15 @@ import org.interledger.ilqp.QuoteByDestinationAmountResponse;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.math.BigInteger;
 import java.time.Duration;
 import java.time.temporal.ChronoUnit;
 import java.util.Objects;
 
 /**
- * An implementation of {@link Codec} that reads and writes instances of
- * {@link QuoteByDestinationAmountResponse}. in OER format.
- * 
+ * An implementation of {@link Codec} that reads and writes instances of {@link
+ * QuoteByDestinationAmountResponse}. in OER format.
+ *
  * @see "https://github.com/interledger/rfcs/blob/master/asn1/InterledgerQuotingProtocol.asn"
  */
 public class QuoteByDestinationAmountResponseOerCodec
@@ -32,9 +33,7 @@ public class QuoteByDestinationAmountResponseOerCodec
     Objects.requireNonNull(inputStream);
 
     /* read the source amount, which is a uint64 */
-    /* NOTE: we don't expect amounts to exceed 2^63 - 1, so we risk the down-cast */
-    // TODO: should we change this?
-    long sourceAmount = context.read(OerUint64.class, inputStream).getValue().longValue();
+    BigInteger sourceAmount = context.read(OerUint64.class, inputStream).getValue();
 
     /* read the source hold duration which is a unit32 */
     long sourceHoldDuration = context.read(OerUint32.class, inputStream).getValue();
