+++ src/main/java/org/interledger/ilp/InterledgerPayment.java
@@ -3,32 +3,42 @@ package org.interledger.ilp;
 import org.interledger.InterledgerAddress;
 import org.interledger.InterledgerPacket;
 
+import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.Objects;
-import javax.money.MonetaryAmount;
 
 /**
- * <p>Interledger Payments moves assets of one party to another that consists of one or more
- * ledger transfers, potentially across multiple ledgers.</p> Interledger Payments have three major
- * consumers: <ul> <li>Connectors utilize the Interledger Address contained in the payment to route
- * the payment.</li> <li>The receiver of a payment uses it to identify the recipient and which
- * condition to fulfill.</li> <li>Interledger sub-protocols utilize custom data encoded in a payment
- * to facilitate sub-protocol operations.</li> </ul> <p> When a sender prepares a transfer to start
- * a payment, the sender attaches an ILP Payment to the transfer, in the memo field if possible. If
- * a ledger does not support attaching the entire ILP Payment to a transfer as a memo, users of that
- * ledger can transmit the ILP Payment using another authenticated messaging channel, but MUST be
- * able to correlate transfers and ILP Payments.</p> <p> When a connector sees an incoming prepared
- * transfer with an ILP Payment, the receiver reads the ILP Payment to confirm the details of the
- * packet. For example, the connector reads the InterledgerAddress of the payment's receiver, and if
- * the connector has a route to the receiver's account, the connector prepares a transfer to
- * continue the payment, and attaches the same ILP Payment to the new transfer. Likewise, the
- * receiver confirms that the amount of the ILP Payment Packet matches the amount actually
- * delivered by the transfer. And finally, the receiver decodes the data portion of the Payment and
- * matches the condition to the payment. The receiver MUST confirm the integrity of the ILP Payment,
- * for example with a hash-based message authentication code (HMAC). If the receiver finds the
- * transfer acceptable, the receiver releases the fulfillment for the transfer, which can be used to
- * execute all prepared transfers that were established prior to the receiver accepting the
- * payment.</p>
+ * <p>Interledger Payments moves assets of one party to another that consists of one or more ledger
+ * transfers, potentially across multiple ledgers.</p>
+ *
+ * <p>Interledger Payments have three major consumers:</p>
+ *   <ul>
+ *     <li>Connectors utilize the Interledger Address contained in the payment to route the
+ * payment.</li>
+ *     <li>The receiver of a payment uses it to identify the recipient and which condition to
+ * fulfill.</li>
+ *     <li>Interledger sub-protocols utilize custom data encoded in a payment to facilitate
+ * sub-protocol operations.</li>
+ *   </ul>
+ *
+ * <p>When a sender prepares a transfer to start a payment, the sender attaches an ILP Payment to
+ * the transfer, in the memo field if possible. If a ledger does not support attaching the entire
+ * ILP Payment to a transfer as a memo, users of that ledger can transmit the ILP Payment using
+ * another authenticated messaging channel, but MUST be able to correlate transfers and ILP
+ * Payments.</p>
+ *
+ * <p>When a connector sees an incoming prepared transfer with an ILP Payment, the receiver reads
+ * the ILP Payment to confirm the details of the packet. For example, the connector reads the
+ * InterledgerAddress of the payment's receiver, and if the connector has a route to the receiver's
+ * account, the connector prepares a transfer to continue the payment, and attaches the same ILP
+ * Payment to the new transfer. Likewise, the receiver confirms that the amount of the ILP Payment
+ * Packet matches the amount actually delivered by the transfer. And finally, the receiver decodes
+ * the data portion of the Payment and matches the condition to the payment.</p>
+ *
+ * <p>The receiver MUST confirm the integrity of the ILP Payment, for example with a hash-based
+ * message authentication code (HMAC). If the receiver finds the transfer acceptable, the receiver
+ * releases the fulfillment for the transfer, which can be used to execute all prepared transfers
+ * that were established prior to the receiver accepting the payment.</p>
  */
 public interface InterledgerPayment extends InterledgerPacket {
 
@@ -53,9 +63,9 @@ public interface InterledgerPayment extends InterledgerPacket {
    * The amount to deliver, in discrete units of the destination ledger's asset type. The scale of
    * the units is determined by the destination ledger's smallest indivisible unit.
    *
-   * @return An instance of {@link MonetaryAmount}.
+   * @return An instance of {@link BigInteger}.
    */
-  Long getDestinationAmount();
+  BigInteger getDestinationAmount();
 
   /**
    * Arbitrary data for the receiver that is set by the transport layer of a payment (for example,
@@ -71,7 +81,7 @@ public interface InterledgerPayment extends InterledgerPacket {
   class Builder {
 
     private InterledgerAddress destinationAccount;
-    private Long destinationAmount;
+    private BigInteger destinationAmount;
     private byte[] data;
 
     /**
@@ -87,9 +97,9 @@ public interface InterledgerPayment extends InterledgerPacket {
     /**
      * Set the destination amount into this builder.
      *
-     * @param destinationAmount An instance of {@link MonetaryAmount}.
+     * @param destinationAmount An instance of {@link BigInteger}.
      */
-    public Builder destinationAmount(final Long destinationAmount) {
+    public Builder destinationAmount(final BigInteger destinationAmount) {
       this.destinationAmount = Objects.requireNonNull(destinationAmount);
       return this;
     }
@@ -119,7 +129,7 @@ public interface InterledgerPayment extends InterledgerPacket {
     private static final class Impl implements InterledgerPayment {
 
       private final InterledgerAddress destinationAccount;
-      private final Long destinationAmount;
+      private final BigInteger destinationAmount;
       private final byte[] data;
 
       /**
@@ -140,7 +150,7 @@ public interface InterledgerPayment extends InterledgerPacket {
       }
 
       @Override
-      public Long getDestinationAmount() {
+      public BigInteger getDestinationAmount() {
         return this.destinationAmount;
       }
